[Refactor/#47] Refactor: 모든 파일에서 `lib/` 폴더 하위로 이동된 경로를 대응
