+++ app/page.tsx
@@ -4,8 +4,8 @@ import Header from "@/components/Header";
 import Search from "@/components/Search";
 import Footer from "@/components/Footer";
 import Button from "@/components/Button";
-import AllPosts from "@/utils/allpost";
-import Post from "@/types/post";
+import AllPosts from "@/lib/utils/allpost";
+import Post from "@/lib/types/post";
 
 const categories: {
   title: string;
